Support different module and function names in models and refactor testnet path adapter
