+++ src/op_analytics/datapipeline/models/compute/execute.py
@@ -1,15 +1,9 @@
-import importlib
 from dataclasses import dataclass, field
-from typing import ClassVar, Protocol
+from typing import Protocol
 
 import duckdb
 
-from .querybuilder import TemplatedSQLQuery, RenderedSQLQuery
-from .types import NamedRelations
-
-
-class ModelFunction(Protocol):
-    def __call__(self, duckdb_client: duckdb.DuckDBPyConnection) -> NamedRelations: ...
+from .model import PythonModel
 
 
 class ModelInputDataReader(Protocol):
@@ -18,49 +12,6 @@ class ModelInputDataReader(Protocol):
         dataset: str,
         first_n_parquet_files: int | None = None,
     ) -> str: ...
-
-
-@dataclass
-class PythonModel:
-    # The registry stores all instances of PythonModel
-    _registry: ClassVar[dict[str, "PythonModel"]] = {}
-
-    # Rendered views caches rendered views so we don't have
-    # to load and render templates each time.
-    _rendered_views: list[RenderedSQLQuery] | None = field(init=False, default=None)
-
-    name: str
-    input_datasets: list[str]
-    expected_output_datasets: list[str]
-    auxiliary_views: list[TemplatedSQLQuery]
-    model_func: ModelFunction
-
-    def __post_init__(self):
-        self._registry[self.name] = self
-
-    @classmethod
-    def get(cls, model_name: str) -> "PythonModel":
-        if model_name not in cls._registry:
-            # Import the module so that the model gets registered.
-            # For now the module and model name are the same by convention, but
-            # that can change if we need to.
-            importlib.import_module(f"op_analytics.datapipeline.models.code.{model_name}")
-
-        return cls._registry[model_name]
-
-    @property
-    def func(self) -> ModelFunction:
-        return self.model_func
-
-    def rendered_views(self) -> list[RenderedSQLQuery]:
-        if self._rendered_views is None:
-            result = []
-            for q in self.auxiliary_views or []:
-                rendered = q.render()
-                result.append(rendered)
-            self._rendered_views = result
-
-        return self._rendered_views
 
 
 @dataclass
